updated vsa-vue
Note:
login.html page has error
(icon tag is missing or not closed)
css not commented properly or mentioned about image use and there link.
Image link can not be permanent so next if you are using img please mention it in readme file or any
so that we converting it will be easy to find out.

Home page same problem , proved proper details of images and other external resource used
<a> tag is not properly  set leading to again reloading of entire page
.